moved fetching functions to utils
